// src/common/api/entities/sys/TypeRefs.ts
export type InvoiceDataGetOut = {
	_type: TypeRef<InvoiceDataGetOut>;

	_format: NumberString;
	address: string;
	country: string;
	date: Date;
	grandTotal: NumberString;
	invoiceId: Id;
	invoiceType: NumberString;
	paymentMethod: NumberString;
	subTotal: NumberString;
	vat: NumberString;
	vatIdNumber: null | string;
	vatRate: NumberString;
	vatType: NumberString;

	items: InvoiceDataItem[];
}
export const InvoiceDataItemTypeRef: TypeRef<InvoiceDataItem> = new TypeRef("sys", "InvoiceDataItem")

export function createInvoiceDataItem(values: StrippedEntity<InvoiceDataItem>): InvoiceDataItem {
	return Object.assign(create(typeModels.InvoiceDataItem, InvoiceDataItemTypeRef), values)
}

export type InvoiceDataItem = {
	_type: TypeRef<InvoiceDataItem>;

	_id: Id;
	amount: NumberString;
	endDate: null | Date;
	itemType: NumberString;
	singlePrice: null | NumberString;
	startDate: null | Date;
	totalPrice: NumberString;
}